moving back searchtype to main file
