--- packages/lib/services/search/types.ts
@@ -1,7 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export enum SearchType {
-	Auto = 'auto',
-	Basic = 'basic',
-	Nonlatin = 'nonlatin',
-	Fts = 'fts',
-}
